backend/pyproject.toml
[project]
name = "talentstreamai-api"
version = "0.1.0"
description = "TalentStreamAI FastAPI service"
requires-python = ">=3.12"
dependencies = [
  "fastapi==0.115.12",
  "uvicorn[standard]==0.34.0",
  "pydantic-settings==2.8.1",
  "python-multipart==0.0.20",
  "httpx==0.28.1",
  "tenacity==9.1.4",
  "PyJWT[crypto]==2.10.1",
  "pypdf==5.4.0",
  "python-docx==1.1.2",
  "boto3==1.40.3",
  "langgraph==1.1.9",
]

[tool.uv]
package = false
